Move start function to lib
